requirements.txt
pandas 
numpy
faker
pyarrow
fastparquet
snowflake
python-dotenv
python-dotenv==1.0.1
generator